increase speed of characters
